Move to Autofac.Core
