+++ src/Autofac/Features/Metadata/MetaRegistrationSource.cs
@@ -23,6 +23,8 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 // OTHER DEALINGS IN THE SOFTWARE.
 
+using Autofac.Core;
+
 namespace Autofac.Features.Metadata
 {
     /// <summary>
